reorganized the add_bus_stops function so that it can be used for external testing
