+++ Scripts/assignment/emme_assignment.py
@@ -4,6 +4,7 @@ from math import log10
 
 import utils.log as log
 from utils.zone_interval import belongs_to_area, faulty_kela_code_nodes
+from utils.network import add_bus_stops
 import parameters.assignment as param
 import parameters.zone as zone_param
 from assignment.abstract_assignment import AssignmentModel
@@ -309,29 +310,8 @@ class EmmeAssignmentModel(AssignmentModel):
 
     def _add_bus_stops(self):
         network = self.mod_scenario.get_network()
-        for line in network.transit_lines():
-            if line.mode.id in param.stop_codes:
-                stop_codes = param.stop_codes[line.mode.id]
-                for segment in line.segments():
-                    is_stop = segment.i_node.data2 in stop_codes
-                    if line.mode.id in "de":
-                        # Non-HSL bus lines
-                        not_hsl = segment.i_node.label not in param.hsl_area
-                        if line.id[-1] == '1':
-                            # Line starts in HSL area
-                            segment.allow_alightings = not_hsl and is_stop
-                            segment.allow_boardings = is_stop
-                        elif line.id[-1] == '2':
-                            # Line ends in HSL area
-                            segment.allow_alightings = is_stop
-                            segment.allow_boardings = not_hsl and is_stop
-                        else:
-                            raise ValueError(
-                                "Unknown direction code for line " + line.id)
-                    else:
-                        segment.allow_alightings = is_stop
-                        segment.allow_boardings = is_stop
-        self.mod_scenario.publish_network(network)
+
+        self.mod_scenario.publish_network(add_bus_stops(network))
 
     def _create_attributes(self, scenario, extra):
         """Create extra attributes needed in assignment.
